Restructured directory in /user/onboarding
