+++ src/app/user/onboarding/_components/Step1.tsx
