+++ src/app/user/onboarding/_components/Step2.tsx
