+++ src/app/user/onboarding/_components/types.ts
